Update IfcWallTypeEnum.md
since STANDARD will be deprecated, there do not seems to be a place for gypsum-block walls and other light weight partitioning walls. PARTITIONING is now described as a light-weight, sandwich construction. I think the massive non-loadbearing walls should be included as well. This is not a change, just an addition to the description.