Update README.md
The horizontal geometry of a Viennese Bend transition curve depends on the corresponding Viennese Bend cant segment. In the special case of more than on child alignment with cant layouts, and those cant layouts have Viennese Bend cant segment with different parameters, the geometry of the horizontal Viennese Bend transition curve is not unique. The geometry of the transition curve segment will depend on which cant segment is used to compute the horizontal geometry.