Amend README.md
The previous README was showing as the follows which are obviously wrong. Changed as per IFC4.2.

Resource Type Assignment

Resource types may have assignments indicating re-usable product types for which occurrences may be sourced by occurrences of the resource type. An example of such assignment is a particular crane model that may be used for a crane equipment resource used for erecting steel.